register

+++ account-service/src/main/java/com/canoe/eb/account/bean/dto/AccountRegisterDTO.java
@@ -10,7 +10,7 @@ public class AccountRegisterDTO {
     @ApiModelProperty(value = "密码", required = true)
     private String password;
     @ApiModelProperty("角色 1->买家  2->商家")
-    private Integer role;
+    private Byte role;
     @ApiModelProperty(value = "验证码", required = true)
     private String captcha;
 
@@ -22,9 +22,9 @@ public class AccountRegisterDTO {
 
     public void setPassword(String password) {this.password = password;}
 
-    public Integer getRole() {return role;}
+    public Byte getRole() {return role;}
 
-    public void setRole(Integer role) {this.role = role;}
+    public void setRole(Byte role) {this.role = role;}
 
     public String getCaptcha() {return captcha;}
